refactor: extract callback function creation in hook
